web/requirements.txt
Flask
Flask-Session
python-dotenv
redis
gunicorn
bcrypt
jwt